fbx2three: Restore ImageUtils, move getDataURL method.
